Connecting user location to Initial Load snapshots for favorites. Connecting distance to UI for favorites.

+++ surf/View/CollectionViews/CollectionViewCells/FavCollectionViewCell.swift
@@ -24,7 +24,7 @@ class FavCollectionViewCell: UICollectionViewCell {
         super.awakeFromNib()
     }
     
-    func setCellContent(waveHeight: Double, waveFrequency: Double, locationName: String, distanceFromUser: Double){
+    func setCellContent(waveHeight: Double, waveFrequency: Double, locationName: String, distanceFromUser: Int){
         heightLabel.text = "\(waveHeight) ft"
         frequencyLabel.text = "\(waveFrequency)s"
         locationLabel.text = locationName
